telegrambot/requirements.txt
python-telegram-bot==20.7
google-cloud-container==2.40.0
google-cloud-billing==1.11.0
google-cloud-resource-manager==1.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
aiohttp==3.9.1
